Extracted helper file for all future assignment creation spec files
